Move topicConsumerStore
