+++ PubSub.Infrastructure.DataAccess/PubSub.Infrastructure.DataAccess.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\PubSub.Core.DomainModel\PubSub.Core.DomainModel.csproj" />
+    <ProjectReference Include="..\PubSub.Core.DomainServices\PubSub.Core.DomainServices.csproj" />
   </ItemGroup>
 
 </Project>
