+++ PubSub.Test/Unit/Application.Services/InMemoryTopicConsumerStoreTest.cs
@@ -2,6 +2,7 @@
 using PubSub.Test.Base.Tests.Toolkit.Patterns;
 using PubSub.Application.Services;
 using PubSub.Core.DomainModel.Consumer;
+using PubSub.Infrastructure.DataAccess.TopicConsumerStore;
 
 namespace PubSub.Test.Unit.Application.Services
 {
